# Enhance Solving Problems - Problem3.cpp - Problem4.cpp - Problem5.cpp - Problem7.cpp - Problem9.cpp - Problem12.cpp - Problem13.cpp - Problem14.cpp - Problem15.cpp - Problem16.cpp - Problem17.cpp - Problem18.cpp - Problem19.cpp

+++ src/_1_problems_from_1_to_10/_1_3_problem_3/Problem3.cpp
@@ -53,7 +53,7 @@ void printSumEachRows(
     short sumEachRows[3],
     const short ROWS
 ) {
-    for (int index = 0; index < ROWS; ++index)
+    for (short index = 0; index < ROWS; ++index)
         cout << "Row " << index + 1 << " Sum = " << sumEachRows[index] << endl;
 }
