+++ src/_1_problems_from_1_to_10/_1_7_problem_7/Problem7.cpp
@@ -34,8 +34,8 @@ void transposeMatrixNumbers(
     const short ROWS,
     const short COLUMNS
 ) {
-    for (int row = 0; row < ROWS; ++row)
-        for (int column = 0; column < COLUMNS; ++column)
+    for (short row = 0; row < ROWS; ++row)
+        for (short column = 0; column < COLUMNS; ++column)
             transposedMatrixNumbers[row][column] = MATRIX_NUMBERS[column][row];
 }
