+++ src/_2_problems_from_11_to_20/_2_2_problem_12/Problem12.cpp
@@ -53,8 +53,8 @@ bool areTypicalMatrices(
     const short ROWS,
     const short COLUMNS
 ) {
-    for (int row = 0; row < ROWS; ++row)
-        for (int column = 0; column < COLUMNS; ++column)
+    for (short row = 0; row < ROWS; ++row)
+        for (short column = 0; column < COLUMNS; ++column)
             if (FIRST_MATRIX_NUMBERS[row][column] != SECOND_MATRIX_NUMBERS[row][column])
                 return false;
     return true;
