+++ src/_2_problems_from_11_to_20/_2_3_problem_13/Problem13.cpp
@@ -22,8 +22,8 @@ bool isIdentityMatrix(
     const short ROWS,
     const short COLUMNS
 ) {
-    for (int row = 0; row < ROWS; ++row)
-        for (int column = 0; column < COLUMNS; ++column) {
+    for (short row = 0; row < ROWS; ++row)
+        for (short column = 0; column < COLUMNS; ++column) {
             if (row == column && MATRIX_NUMBERS[row][column] != 1)
                 return false;
             if (row != column && MATRIX_NUMBERS[row][column] != 0)
