+++ src/_2_problems_from_11_to_20/_2_4_problem_14/Problem14.cpp
@@ -23,8 +23,8 @@ bool isScalarMatrix(
     const short COLUMNS
 ) {
     const short TARGET_NUMBER = MATRIX_NUMBERS[0][0];
-    for (int row = 0; row < ROWS; ++row)
-        for (int column = 0; column < COLUMNS; ++column) {
+    for (short row = 0; row < ROWS; ++row)
+        for (short column = 0; column < COLUMNS; ++column) {
             if (row == column && MATRIX_NUMBERS[row][column] != TARGET_NUMBER)
                 return false;
             if (row != column && MATRIX_NUMBERS[row][column] != 0)
