+++ src/_2_problems_from_11_to_20/_2_7_problem_17/Problem17.cpp
@@ -60,8 +60,8 @@ bool isTargetNumberFoundInMatrixNumbers(
     const short COLUMNS,
     const short TARGET_NUMBER
 ) {
-    for (int row = 0; row < ROWS; ++row)
-        for (int column = 0; column < COLUMNS; ++column)
+    for (short row = 0; row < ROWS; ++row)
+        for (short column = 0; column < COLUMNS; ++column)
             if (MATRIX_NUMBERS[row][column] == TARGET_NUMBER)
                 return true;
     return false;
