+++ src/_2_problems_from_11_to_20/_2_8_problem_18/Problem18.cpp
@@ -41,8 +41,8 @@ void printIntersectedNumbers(
     const short ROWS,
     const short COLUMNS
 ) {
-    for (int row = 0; row < ROWS; ++row)
-        for (int column = 0; column < COLUMNS; ++column)
+    for (short row = 0; row < ROWS; ++row)
+        for (short column = 0; column < COLUMNS; ++column)
             if (
                 const short FIRST_NUMBER = FIRST_MATRIX_NUMBERS[row][column],
                             SECOND_NUMBER = SECOND_MATRIX_NUMBERS[row][column];
